renaming LksApiUtils.ts to LksgUpload.ts
