+++ dataland-frontend/tests/e2e/specs/frameworks/FrameworkSelection.ts
@@ -5,7 +5,7 @@ import { FixtureData } from "@e2e/fixtures/FixtureUtils";
 import { LksgData } from "@clients/backend";
 import { uploadOneEuTaxonomyFinancialsDatasetViaApi } from "@e2e/utils/EuTaxonomyFinancialsUpload";
 import { generateEuTaxonomyDataForFinancials } from "@e2e/fixtures/eutaxonomy/financials/EuTaxonomyDataForFinancialsFixtures";
-import { uploadCompanyAndLksgDataViaApi } from "@e2e/utils/LksgApiUtils";
+import { uploadCompanyAndLksgDataViaApi } from "../../utils/LksgUpload";
 import { MEDIUM_TIMEOUT_IN_MS } from "../../utils/Constants";
 import { getPreparedFixture } from "../../utils/GeneralApiUtils";
